Fix some bugs

+++ GpsTracker.Service/GpsTrackerService/GpsTracker.Models/Messages/SosMessage.cs
@@ -11,8 +11,8 @@ namespace GpsTracker.Models.Messages
         public int UserId { get; set; }
         //public string Login { get; set; }
         //public string Password { get; set; }
-        public float Latitude { get; set; }
-        public float Longitude { get; set; }
+        public double Latitude { get; set; }
+        public double Longitude { get; set; }
         public DateTime Timestamp { get; set; }
     }
 }
